refactor: add example module and rename core to ckb and rename console to ckb-sdk
