+++ ckb-sdk/build.gradle
@@ -1,11 +1,4 @@
 description 'ckb-sdk-java is a lightweight Java library for integration with nervos CKB.'
 
-dependencies {
-    compile project(":core")
-    compile project(":utils")
-    compile "com.google.code.gson:gson:$gsonVersion"
-    compile "org.slf4j:slf4j-api:$slf4jVersion"
-}
-
 apply from: rootProject.file('gradle/checkstyle.gradle')
 apply plugin: 'com.github.johnrengelman.shadow'
