+++ ckb/src/main/java/org/nervos/ckb/type/BannedAddress.java
