+++ ckb/src/main/java/org/nervos/ckb/type/BannedResultAddress.java
