+++ ckb/src/main/java/org/nervos/ckb/type/CellbaseOutputCapacity.java
