+++ ckb/src/test/java/type/ScriptTest.java
